Dynamics

Dynamics

ED_sectors.evolve_state — Function.

evolve_state(psi, eigsys, t)

A function which computes the time-evolution of a state psi under a (already diagonalized) Hamiltonian eigsys for time t. Explicitly, this computes $\left.e^{- i H t} |\psi\right>$.

source

ED_sectors.expectation_time — Function.

expectation_time(O, t, psi, eigsys)
expectation_time(O, t, rho, eigsys

For a state psi, measures the expectation of the operator O at time t evolved under eigsys: $\left<\psi|O(t)|\psi\right>$.

For a density operator rho, measures the expectation of the operator O at time t evolved under eigsys: $\operatorname{Tr}[\rho\; O(t)]$.

source

ED_sectors.autocorrelation — Function.

autocorrelation(O, t, rho, eigsys)

Computes the autocorrelation of an operator O at time t against the density matrix rho when evolved under eigsys: $\operatorname{Tr}[\rho\; O(t) O(0)]$.

source