Full Spectrum Measurements

Full Spectrum Measurements

ED_sectors.EIGSYS — Type.

EIGSYS(evs, U)
EIGSYS(H)

A struct for storing the full spectrum evs and eigenstates of a system U. Explicitly, U is the (unitary) matrix which gives the change of basis from the original to energy bases, i.e. the columns of U are the eigenvectors of the system. So for an EIGSYS ES, the full spectrum is ES.evs and the 3rd eigenvector, for example, is ES.U[:,3].

Examples

julia> L = 4; basis = BASIS(L);

julia> ising = ABSTRACT_OP(L; name="Ising Model", pbc=true) + 2TERM("ZZ") + TERM("X");

julia> H = Operator(ising,basis);

julia> ES = EIGSYS(H)
EIGSYS[dim: 16]

julia> psi3 = ES.U[:,3]
16-element Array{Complex{Float64},1}:
  0.08727992975105786 + 0.0im
 -0.23235256582864727 + 0.0im
  -0.2323525658286471 + 0.0im
   0.3509044118546869 + 0.0im
 -0.23235256582864688 + 0.0im
 -0.17367654405779392 + 0.0im
  0.35090441185468646 + 0.0im
 -0.23235256582864808 + 0.0im
 -0.23235256582864716 + 0.0im
   0.3509044118546871 + 0.0im
  -0.1736765440577938 + 0.0im
 -0.23235256582864866 + 0.0im
  0.35090441185468646 + 0.0im
 -0.23235256582864844 + 0.0im
 -0.23235256582864783 + 0.0im
  0.08727992975105792 - 0.0im

julia> isapprox((H * psi3),(ES.evs[3]*psi3)) #they are the same to numerical error
true

source

ED_sectors.thermal_density_matrix — Function.

thermal_density_matrix(eigsys, beta)

Computes the thermal density matrix in the energy basis for a given eigsys and inverse temperature beta. When beta = 0, this gives the "infinite-temperature" density matrix.

Warning

It is easy to get underflow or overflow errors by exponentiating large numbers. The largest Float64 is roughly $10^{308} \sim e^{709}$. It is recommended to double-check that your results make sense after using this function.

source

ED_sectors.r_statistic — Function.

r_statistic(eigenvalues :: Array{Float64})
r_statistic(ES :: EIGSYS)

Computes the R-statistic for a spectrum. The R-statistic is a measure of how close to integrability the system is. For integrable systems, $\left<r\right> \sim 0.386$. For GOE level-statistics, $\left<r\right> \sim 0.5295$.

Warning

Using the R statistic correctly is subtle. It only makes sense systems sufficiently large as to be close to the thermodynamic limit. Moreover, one must restrict to a single symmetry sector. Any residual symmetry will give an artificially low R-statistic. See, e.g., https://arxiv.org/abs/cond-mat/0610854 for details..

source